requirements.txt
langchain
langgraph
langchain-community
langchain-openai
chromadb
python-dotenv
tiktoken
langchain-tavily
sentence-transformers
unstructured
pypdf
openpyxl
python-magic-bin
pandas
msoffcrypto-tool
nltk
langchain-huggingface
tqdm
langchain-chroma
